[REMOVE] 회원가입 성별,연도 페이지 삭제(#23)
