+++ src/pages/SignUp/SignUp2.tsx
@@ -1,61 +1,86 @@
-import React, { useState } from "react";
-import { useNavigate } from "react-router-dom"; // 페이지 이동을 위한 훅
-import ProgressBar from "../../components/ProgressBar";
+import React, { useEffect, useState } from "react";
+import { useLocation, useNavigate } from "react-router-dom";
 
 const SignUp2: React.FC = () => {
+  const location = useLocation();
   const navigate = useNavigate();
-  const [age, setAge] = useState(""); // 나이 상태 관리
+  const { nickname = "회원" } = location.state || {};
 
-  const handleAgeChange = (event: React.ChangeEvent<HTMLInputElement>) => {
-    setAge(event.target.value); // 입력값 업데이트
-  };
+  const [progress, setProgress] = useState(0);
 
-  const handleAgeSubmit = () => {
-    console.log(`Entered Age: ${age}`);
-    navigate("/signup3"); // SignUp3 페이지로 이동
-  };
+  useEffect(() => {
+    let currentProgress = 0;
+    const interval = setInterval(() => {
+      currentProgress += 1;
+      setProgress(currentProgress);
 
-  const handleKeyPress = (event: React.KeyboardEvent<HTMLInputElement>) => {
-    if (event.key === "Enter") {
-      handleAgeSubmit(); // 엔터 키 입력 시 다음 페이지로 이동
-    }
-  };
+      if (currentProgress >= 100) {
+        clearInterval(interval);
+        setTimeout(() => {
+          navigate("/signup3");
+        }, 500); 
+      }
+    }, 30);
 
-  const handleBlur = () => {
-    if (age) {
-      handleAgeSubmit(); // 입력 필드에서 포커스가 벗어나면 이동
+    return () => {
+      clearInterval(interval);
+    };
+  }, [navigate]);
+
+  const getCircleState = (step: number) => {
+    if (progress >= step * 33.33) {
+      return "bg-blue-500"; 
+    } else {
+      return "bg-gray-300"; 
     }
   };
 
   return (
     <div className="h-screen flex flex-col items-center bg-blue-50 font-pretendard px-6">
-      {/* Progress Bar */}
-      <div className="w-full max-w-md mx-auto mt-16">
-        <ProgressBar progress={50} />
+      {/* Circular Progress */}
+      <div className="mt-36 relative w-[148px] h-[148px] flex items-center justify-center">
+        <img
+          src="/Ellipse14.png"
+          alt="Circular Progress"
+          className="w-full h-full"
+        />
+        {/* 중앙 숫자 */}
+        <span className="absolute text-black text-[16px] font-bold">
+          {progress}
+        </span>
       </div>
 
       {/* Title */}
-      <div className="text-center mt-6 w-full max-w-md mx-auto">
-        <h1 className="text-[40px] font-[GmarketSansWeight] font-bold text-black ml-4 text-left whitespace-pre-line">
-          출생 연도를{"\n"}입력해 주세요.
+      <div className="text-center mt-8 w-full max-w-md">
+        <h1 className="font-[GmarketSansWeight] text-[32px] font-[GmarketSansWeight] text-black whitespace-pre-line leading-9">
+          {nickname}님을 위한 식단을{"\n"}추천하는 중입니다
         </h1>
-        <p className="font-semibold font-[pretendard] mt-8 ml-4 text-left text-[16px]">
-          나이에 따라 하루 에너지가 달라져요!
-        </p>
       </div>
 
-      {/* Input Section */}
-      <div className="w-full mx-auto max-w-xs mt-4">
-        <input
-          type="text"
-          id="age-input"
-          placeholder="예: 2001년"
-          className="w-[327px] h-[55px] bg-white font-[pretendard] border border-[#EDEDED] shadow-whiteBox rounded-base px-4 text-[14px] text-gray-800 placeholder-gray-400 outline-none"
-          value={age}
-          onChange={handleAgeChange} // 입력값 업데이트
-          onKeyPress={handleKeyPress} // 엔터 키 이벤트 처리
-          onBlur={handleBlur} // 포커스 아웃 이벤트 처리
-        />
+      <div className="mt-12 ml-28 flex flex-col gap-6 w-full max-w-md">
+        {/* 첫 번째 항목 */}
+        <div className="flex items-center gap-3">
+          <div className={`w-4 h-4 rounded-full ${getCircleState(1)}`}></div>
+          <p className="font-[pretendard] text-gray-700 text-[14px] font-semibold">
+            중량을 위한 효율적인 방법 계산 중
+          </p>
+        </div>
+
+        {/* 두 번째 항목 */}
+        <div className="flex items-center gap-3">
+          <div className={`w-4 h-4 rounded-full ${getCircleState(2)}`}></div>
+          <p className="font-[pretendard] text-gray-700 text-[14px] font-semibold">
+            하루 목표 칼로리 계산 중
+          </p>
+        </div>
+
+        {/* 세 번째 항목 */}
+        <div className="flex items-center gap-3">
+          <div className={`w-4 h-4 rounded-full ${getCircleState(3)}`}></div>
+          <p className="font-[pretendard] text-gray-700 text-[14px] font-semibold">
+            건강을 위한 영양 지표 도출 중
+          </p>
+        </div>
       </div>
     </div>
   );
